Update on "[quant][graphmode][fix] Fold conv bn"
Summary:
1. applied filter for the module types
2. removed the assumption that the conv bn are immediate child of parent module

Test Plan:
python test/test_quantization.py TestQuantizeJitPasses
Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]

+++ test/quantization/test_quantize_jit.py
@@ -1097,7 +1097,7 @@ class TestQuantizeJitPasses(QuantizationTestCase):
 
     def test_inplace_option(self):
         for tracing in [True, False]:
-            model = get_script_module(torch.nn.Conv2d(3, 3, 3), tracing, self.img_data[0][0])
+            model = get_script_module(torch.nn.Conv2d(3, 3, 3).float(), tracing, self.img_data[0][0])
             qconfig_dict = {'': default_qconfig}
             quantize_jit(
                 model, qconfig_dict, test_only_eval_fn, [self.img_data], inplace=True)
